+++ pkg/core/packages.go
@@ -273,7 +273,7 @@ func (s PkgDeplSpec) ResAttachmentSource(parentSource []string) []string {
 // DefinesApp determines whether the PkgDeplSpec instance defines a Docker Compose app to be
 // deployed.
 func (s PkgDeplSpec) DefinesApp() bool {
-	return len(s.DefinitionFiles) > 0 && s.DefinitionFiles[0] != ""
+	return len(s.ComposeFiles) > 0 && s.ComposeFiles[0] != ""
 }
 
 // PkgFeatureSpec